Refactor code to extend KFP SDK compiler (#16)
This code change removes the requirement to clone and
copy the kubeflow/pipelines repository. The project
has been refactored to only contain the minimal set of
Python classes and methods that needed to be modified
to generate Tekton YAML instead of Argo YAML. The
overriden methods are dynamically linked at runtime
by monkey-patching the KFP SDK module.

Additionally adding a simple test case and setup.py
install script that creates the dsl-compile-tekton
CLI executable which works independent of the KFP
compiler executable.