Remove build.sh in favor of make distribution (#213)
Using the build.sh script is dangerous as it does not ensure
the safeguard of a virtual environment. It modifies the users
current Python environment, which could be the system Python
environment. This could cause problems with incompatible
Python packages, and it can be tedious at best to restore
the original Python environment.