Update workflow.go (#794)
* Update workflow.go

In verifyParameters, check if the parameter is on the list of unused parameters, and if so --- log a warning instead of raising a hard error.
If there are too many parameters, it is hard to determine which one is in use, which one is unused, so here just a warning will be better for this case.

* Update workflow.go

* Update workflow.go